+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
  - also return and store uncertainty estimates (stds) in de analysis
  - also return and store z-scores in de analysis
  - implement underrepresentation filtering for de analaysis
+ - `plot.embedding` scanpy wrapper can now plot multiple layer
 
 ## [0.3.3]